Add a generalized ui element module
The idea is the lists of sliders, buttons and gauges and the displaymode
crap can all be put into this generalized ui element stuff and removed
from the specific ui element code.  This commit just adds code for the
generalized ui elements, but does not yet fix the individual sliders, buttons
and gauges yet.