Fix issue with ssgl leaking file descriptors
After awhile, snis_server would complain about accept()
failing due to too many open files.  I think this should
fix that.  I think what was happening was that as the server
tried to update ssgl_server, the write failed because the ssgl
server gave up on it (this may be a separate problem) and then
it leaked the descriptor.