fix bug in initial updates for clients
Some objects rarely or never update (e.g. nebula).  There is an optimization
that far-away objects sometimes skip updates.  When a client first connects, if
he misses the initial update for a nebula or other such object due to this
optimization, he will never get an update for that object and will not know it
even exists.  Do not do this far-away-update-skip optimization for clients that
just joined.