remove sdata requesting.
Science station would send requests to server for "sdata" for things
it was scanning.   This was silly.  The server already knows everything
about the science station and could know which things would be requested.
So might as well eliminate the requesting mechanism and just have the
server send the sdata as appropriate without being requested.