only change screen submodes if mainscreen supported or if you are the requesting client.
Jeremy doesn't like that the sub-modes (e.g. nav 2D and nav 3D, the different
science screen modes) are ship-wide and thinks they should be per client.  But
making them purely per-client breaks the idea of Picard being able to say "onscreen!"
and have the screen magically appear on the main screen via either the station
in question pressing ctrl-O, or the comms officer making it happen.

So we arrived at a compromise that if a station supports main screen, then it
will get requests for changing the sub modes, or, of course the station which
requests such a change will get those requests from the server, but stations
other than the requesting station which do not support main screen will not get
those requests.

In this way, multiple people operating science stations (spectators) will be
able to choose different modes independently so long as their stations do not
have the MAIN_SCREEN_ROLE, and those stations that do have the MAIN SCREEN role
will still get those requests, and so will be able to mirror the various
stations.