improve initial position interpolation for newly added objects
We know the velocity, we can calculate the "previous" position by
applying that negated velocity, then we use that "previous" position for
position interpolation immediately without waiting for a second update
to begin the interpolation.

This makes a noticeable difference with the laser bolts.  Without this,
they kind of just sit there until the 2nd update comes (usually within
1/10th second) then they take off, but it's noticeable and weird looking,
esp. if there is a delay for that 2nd update.