Add a texture mapping shader and use it for laser beams
There are a few things wrong in here, e.g. texture number is
stored in the shader structure, should be in the mesh structure,
and the vertex shader needs some cleaning up, and the per-pixel-
lighting shaders don't exist.

But, it's kind of working, and compared to what it was like before
at least, it looks badass, so I'm going to commit it.