fix bug in processing removal of torpedoes and lasers
In the "move" functions if the torpedo or laser is not alive
it should delete it self and nothing more (e.g. do not do one
last collision detection.)  Otherwise, items spawned as the result
of destruction (e.g. cargo containers popping out of a ship) immediately
will get hit by the same torpedo or laser that caused them to exist,
and immediately get destroyed.