fix ghost ship issue
Was using player_collision_detection when moving NPC ships.  NPC
ships would crash into planets, which would kill the NPC ship, but
do so in the manner of the player ship -- that is, not remove it,
on the presumption that it would respawn.  But it doesn't respawn.
So that left the server with a dead ship that the clients didn't
know was dead because the server doesn't send updates for dead
things.

So, don't do that.  Still need some sort of collisions detection
thing for NPC ships.