gaseous-giganticus: more trying to get rid of velocity divergences
I think it has to do with how we decided whether or not a noise gradient is
uphill or downhill.  Put in a 0.95 fudge factor, which improved things a lot,
but do not know how to figure the "right" answer (or even if there *is* a right
answer that avoids divergences.