Read exhaust port data at runtime instead of building into binary
We process scad files to produce header files containing exhaust port
locations and scaling for spaceship models either within scad files or
manually.  These header files are then baked into the binary.  This is
at odds with the goal of having the models be replaceable independent of
the binary.  So, we should read this data at runtime, just as we read the
other parts of the model data (.stl and .obj files) at runtime.