Improve tesselation of spherified cube by intelligently slicing rectangles
The 6 faces of the cube we start with are divided into a number ofi rectangles,
and each rectangle is divided into two triangles.  When dividing into triangles,
we should choose the shortest diagonal of the rectangle to produce the largest
triangles that are closer to being equilateral triangles and avoid unnecessarily
creating triangles that are thinner slivers than they need to be.

Signed-off-by: Stephn M. Cameron <stephenmcameron@gmail.com>