Fix problem with main screen button on comms
From the bug report:

"here a bunch of buttons at the top of the comms screen that set the main
screen, There is a button called "Main" which does nothing. The other buttons
throw the requested display on the main screen. To return to the
out-of-the-window view you have to press one of the other buttons again, like
it's a toggle."

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>