Fix SCI_SELECT_COORDS opcode format discrepancy
This was wrong for a long time with client transmitting "bww"
but server thinking it should be "bSS" (uint32s vs. 32-bit encoded
floating point numbers). Since they were the same size, and since
this opcode is really obsolete (left over from the 2D days), it didn't
manifest as anything until I added the opcode checking code, at which
point it is detected.  The effect was randomly clicking on the science
screen would cause the game to detect a problem in the protocol
and seem to freeze up.  This should fix that.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>