Fix unwanted warp tunnel rendering on switching to main view.
The way it worked was to store the ship's last position on each
frame, then do vector subtraction to calculate a velocity to determine
whether to show the warp tunnel.  If you switch away from the main
screen, travel some distance, then switch back, it thinks you're going
very fast, and shows the warp tunnel for a frame.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>