Suppress broken semi-autonomous mode of damage control robot
Clicking with the mouse can set up short or long term goals for
the robot.  This was originally used for testing the autonomous
mode.  When you click with the left or middle mouse button it set
up either a short term or long term destination for the robot, which
would then try to figure out how to get there.  However, this conflicts
with the (little used) buttons on the damcon screen. Clicking those buttons
(e.g. left, right, forward, backward) ends up setting up goals for
the robot to regions inside the buttons which the robot cannot reach,
and also takes the robot out of manual mode.

Since the manual mode works, and the autonomous mode works, and the
semi-autonomous mode (setting short and long term goals with mouse
clicks) is broken... let's just disable the semi-autonomous mode.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>