Fix bug in setting button focus on comms screen
ui_set_focus takes a struct ui_element *, not a widget pointer.
Have to lookup the ui element from the widget pointer first.

Also, make buttons set focus to themself *before* calling the
button callback if inside_fn is set (which it almost always is,
the exceptions could be considered to be minor bugs.) This is in
case the callback wants to set the focus itself to something else
(e.g. to the comms input field). The button press should not undo
that by resetting the focus to the button after the button callback
has changed it to comms input field.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>