Add (completely untested) DMX lighting library
This code almost certainly does not work.  In fact, I have not yet
even ever executed any of this code.  I currently have no hardware to
test it with, and I wrote it with only a rudimentary understanding
of how the hardware works based on a few things I read on the
internet. Maybe one day, this will eventually work, but for now,
it almost certainly doesn't. That being said, snis_dmx.o is not
linked into any executables, so it shouldn't cause any harm if you do
not execute it.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>