Allow solarsystems to have associated random seeds.
Previously, solarsystems had random seeds associated with them
via the SNISRAND variable and the "quickstart" script. Now the
random seed is stored in the solarsystem config, though it can
still be overridden by the SNISRAND environment variable. Now
the quickstart script does not use SNISRAND. Previously the
"snis_launcher" script also used SNISRAND, but not in the same
way as "quickstart", so the association between random seeds and
solarsystems was inconsistent. That's fixed now.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>