Add a new_bootable_media_instructions.html for latest Linux Mint, and Linux Live Kit to gh-pages site.
I found that mintconstructor is no longer available to the public, and also old
instructions were for a pretty old version of Linux Mint. Update includes 18.3
and uses Linux Live Kit instead of minconstructor. New page is not yet linked
to, I can throw in an extra commit if you'd like that.

Also, new page is laid out in more "html5 standard" way and adds a CSS file.
This may be useful as a basis for a page template for the rest of the site.

(Some very minor whitespace clean-ups on this patch by Stephen M. Cameron)

Author: Justin Warwick <justin.warwick@gmail.com>
Signed-off-by: Justin Warwick <justin.warwick@gmail.com>
Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>