Fix inconsitencies in comparisons involving snis_object_pool_highest_object()
The value it returns is the valid index of the highest extant object, so
comparisons of the form:

if (i >= snis_object_pool_highest_object(pool))
	printf("i is not valid\n");

are incorrect.

Similarly, for loops should be:

	for (i = 0; i <= snis_object_pool_highest_object(pool); i++)

rather than,

	for (i = 0; i < snis_object_pool_highest_object(pool); i++)

There were several instances of the latter incorrect form in snis_server.c,
mostly in later additions in the RTS code done after I had apparently forgotten how
it worked.

Thanks to Justin Warwick for discovering this problem.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>