Fix problem with mainscreen weapons view mode
Shift-W is used to toggle the main screen between weapons view and normal view.
This can be triggered from Nav, Weapons, and Main screen. However, if the client
does not have main screen role, and triggers weapons mode from Nav screen, the
main screen clients will switch, but the non-mainscreen client will not be able
to toggle weapons view off. This is because the server sends the view mode update
only to clients that have main screen role. It should also update clients with
Nav and Weapons roles.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>