Print a message if client doesn't get ship ID from server after awhile.
In main_da_expose(), when you can't find the player ship pointer after
4 seconds of trying (e.g. server is slow, or never sends it for some reason), do
not just sit there with a blank screen, put out a message.

This transforms the "black screen" problem into the "CANNOT FIND POINTER TO
PLAYER SHIP, WAITING" problem, which is a bit more informative, though no less
annoying.

The problem is somewhat reproducible. If I start 5 or 6 clients usually it will
happen. If I only start 1 or 2 clients it pretty much never happens. So something
is wrong on the server side perhaps. Either it's not sending client ID packet, or
the client isn't getting that packet for some reason. Still debugging it, but at
least this message is better than a black screen.)

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>