Hide lobby button when IP addr is 255.255.255.255
If you run 2 clients on the same host, then the 2nd client cannot
recieve the lobby broadcast packets, so it won't find the lobby IP
address. Probably should use multicast instead of broadcast to get
around that.

I thought I already hid the button if IP addr is 255.255.255.255, but
apparently I left some debug code in that suppressed that behavior.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>