Fix problem of crashing when quit screen is active too long
I noticed that if you press ESC to enter the quit dialog, then
just sit there for awhile without doing anything, eventually the
program crashes.  The screen doesn't update because we returned
from advance game early, but the server is still ticking, so things
are still moving. If we don't return early, the crash doesn't happen,
and the screen keeps updating.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>