Fix snis_client crash on invalid my_ship_oid
After a warp gate jump, my_ship_id gets set to a new value,
and my_ship_oid gets reset to -1, and there is a period of time
when the corresponding go[] entry for my_ship_id does not exist.
This means that there are times when accessing go[my_ship_oid]
can crash. We need to be extra careful when using my_ship_oid.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>