Allow locking all client roles from the demon screen
At a recent play session, one player suggested that it would be nice
if players could be limited to their assigned roles.  Of course this is
possible, either by selecting the roles at login, or by using the "role"
command from the demon screen to manually lock down the roles of each
client.  However, the former requires foresight, and the latter is too
cumbersome. It occurred to me that it is possible to simply command each
client to lock down to whatever is the current role, and this would
accomplish the player's wish.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>