Take angle of incidence into account for specular reflectivity of water
If the light is coming straight down on the water, most of it gets transmitted into
the water and very little is reflected.  If the light is coming onto the water at a
low glancing angle, then most of it is reflected, and very little is transmitted. Our
planet shader should take this effect into account (at least with a very very loose
approximation anyway).

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>