Change main camera near plane from 1.0 to 5.0.
This makes the boundary between rendering passes move from 10000 to 50000 units,
which in turn means that the boundary is quite far away, and so does not produce
quite such obnoxiously terrible artifacts on planets.  10000 was a particularly bad
distance for this boundary to be at, because planets would typically intersect this
boundary whenever you were near them, and the artifacts were particularly bad with
the atmosphere and planet surface.

Interestingly, trying to make the artifact occur with planets at a distance of 50000
units, it seems I cannot now make it occur.  Quite surprisingly this change seems to have
TOTALLY and COMPLETELY fixed it!  Holy crap, I'm glad to see this artifact reduced.
I did not quite realize how much it bothered me until it was gone.

Empirically, running through objects (e.g. starbases) does not produce objectionable
artifacts due to the changed near plane value of 5, as this is still quite close to
the camera.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>