Choose farthest rendering pass boundary with fewest intersections.
Previously, if all the rendering pass boundary candidates had an equal
number of object intersections, the nearest one would be chosen. We should
instead choose the farthest, so that any resulting artifacts are likely to
occupy less screen space and so be less obnoxious, or with a little luck
are even made completely invisible due to being moved to, e.g. the back
side of a planet.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>