Transmit science selection to NAV and WEAPONS in addition to SCIENCE
This partially fixes the problem that the science selection is not transmitted to all
roles that need it.  If you have two clients with restricted roles, e.g. one client with
SCIENCE, TTS, SOUND, and DEMON roles, and another with NAV, TTS, SOUND, and DEMON roles,
if you select something on SCIENCE, you would expect to see the SCIENCE arrow on
the NAV console, however, it did not appear. This is because the science selection was
only transmitted to clients with role SCIENCE.

We still have a bug though, because the science selection should really be transmitted
as part of updating the player ship rather than at the time of selection, because if we
transmit only at the time of selection then we do not handle the case of consoles that
join the game after the science selection has already been made.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>