Allow IP address probe address to be specified via environment variable
snis_server needs to know "the IP address" of the host it runs on so it
can communicate this to ssgl_lobby.  This is more complicated than it seems,
because a host may have multiple IP addresses, and the lobby host might be
accessible via localhost, but telling the lobby that the IP address is
127.0.0.1 is not useful, so we need a way to get the "primary" IP address.
So we do that by constructing a datagram socket connected to 8.8.8.8,
Google's DNS server, then using getsockname to figure out the IP address
(network interface) that our host would be use to transmit, if we did
transmit (which we don't).  So far so good.

However, some people want to run SNIS on a private network that does not
have any connection to the internet.  In that case, connect(2) fails, and
ssgl_get_primary_host_ip_addr() returns 0.0.0.0, which is duly registered
with the lobby, but is useless for actually connecting.  So to accomodate
this scenario, we need to allow users to specify a different probe address
to determine the primary network interface, e.g. 192.168.1.1 for example.

So we introduce a new environment variable, for example:

export SSGL_PRIMARY_IP_PROBE_ADDR=192.168.1.1

Then run snis_launcher, and everything will use this probe addr for
determining the IP address, and connect(2) will succeed even if there's
no internet connection, and all will be well.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>