Avoid undefined behavior when fixing up internal pointers after realloc()
When realloc()ing the vertices in a mesh (realloc m->v[]) to add more vertices,
all the vertex pointers in the triangles in m->t[].v[] need to be fixed up
(they all point into m->v[]).

Previously we did this by saving the old value of m->v, and then subtracting
the old value and adding the new value.  However, using the old value in any
way after realloc has freed it is apparently undefined behavior. And clang
scan-build would complain about it as a "use after free" (even though we
never dereference the old pointer value.) Empirically this worked everywhere
I tried it (x86, and x86_64).

So to avoid the undefined behavior, before the realloc, we convert all those internal
pointers into integer offsets which we store in the space used for the pointers as intptr_t's.
Then after the realloc, we convert them back into pointers.  This avoids using the old
pointer value after it has been freed, at the cost of more or less doubling the
amount of work that we need to do.  Luckily it's not in a performance critical
area.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>