Fix potential use-after-free in text_to_speech_thread_fn()
I don't think this is a real problem, but the fix is easy and makes sense.

In order for this to be a real problem, text_to_speech_wait_for_input() would
have to return with text_to_speech_thread_time_to_quit == FALSE and with
text_to_speech_queue_head == NULL, which is impossible, AFAICT.  However,
setting entry to NULL after freeing it means clang scan-build does not need
to be smart enough to realize the impossibility in order not to complain.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>