Try to detect and stop runaway lua scripts
This works by using lua_sethook with the LUA_MASKCOUNT parameter to set
a call back after some number of Lua VM instructions have been executed.
That number is by default 100000, but is tweakable up to 1000000.

This isn't perfect.  If a Lua script calls a C function and that C function
enters an infinite loop, this method will not help.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>