Allow bandwidth throttling measures to be tweaked
Also reduce distance at which updates are throttled from 300000 to 200000
and increase distant object update period from 0.7 secs to 2.0 secs.
This reduces bandwidth usage by more than half from ~130k/s to ~55k/s
per client at the cost that distant objects are updated less frequently
and that "distant" is a little closer than it used to be. The values are
now tweakable, so you can set DISTANT_UPDATE_PERIOD to 7 and
BANDWIDTH_THROTTLE_DISTANCE to 300000 to get the behavior prior to this
patch.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>