Fix wiggling NPC ships
The problem appears to be with very slow moving ships that are very near
their destinations. To avoid this wiggling, we suppress orientation changes of
very slow moving ships.  This is all a bit finicky.  If the threshold velocity
at which we suppress orientation changes is too high, then ships do not wiggle,
but have a hard time getting close to their destinations.  If too low, then
wiggling occurs.

This change seems to improve things in that it reduces or eliminates wiggling
while allowing ships to reach their destinations reasonably well.  This change
may not be the last word on this topic though. I still don't *really* understand
the wiggling.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>