Fix problem with extra quotes on solarsystem name
This was causing a race with starting a solarsystem manually
via snis_launcher and multiverse autowrangler starting a snis_server
instance automatically. The manually created solarsystem was name
"DEFAULT", with quotes, and the automatically created one was called
DEFAULT without the quotes.  Autowrangler would then not recognize
that these were supposed to be the same.  Or something along those
lines.  In any case, the quotes weren't supposed to be there.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>