Remove initialize.lua script and code that calls it
initialize.lua isn't really used for anything, and is only called
when snis_server starts. Notably, it is not called after clear_all(),
so any callbacks it might have registered will be wiped out when
clear_all() is called (as is commonly done at the beginning of mission
scripts.)  We could cause clear_all() to re-run initialized.lua
immediately after everything is clear to re-instate whatever
callbacks were set up by initialize.lua, and this might make
initialize.lua potentially more useful.  But is is only potentially
more useful, not actually, as since its inception on 2013-09-14
initialize.lua has not changed, and the only things in it are some
callbacks that just print out strings for a few events, essentially
test code dating back to when the event callback system was created.

So, probably best to just nuke it.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>