Document mysterious multiplication by an angle
in process_vertex_normals() in stl_parser.c, there is a mysterious multiplication
by an angle.  The idea is as you are figuring how much each triangle adjacent
to a vertex should contribute to a vertex normal, you weight them by the angle
they form at that vertex, so triangles with large angles contribute more to
the vertex normal and triangles with small angles contribute less.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>