Separate pull down submenu width from submenu title width
On low res screens (720p) the pull down menu on the demon screen
was messed up.  It calculated each submenu width based on the widest
thing in the submenu.  When you add all those up, they are wider than
the screen, which pushed the leftmost submenu off the screen.  The
submenu items width should not impact the submenu title width.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>