Fix twitchy ETA values in science.
When orbiting a planet, if you target sensors on the planet,
the ETA would show strange values as your course was purposely
and always nearly perpendicular to the direction towards the
planet.  The threshold for displaying the ETA needs to be a bit
further from 0 than 0.001.  Changed it to 2.0, arrived at
empirically while orbiting a selected planet.

Also, use the dejittered range and closing rate values to
calculate the ETA.

Also added a slight amount of hysteresis to range and closing
rate to prevent the values from crossing back and forth over
the threshold too frequently to cut down on the ETA flickering
on and off too much.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>