Unify the thread creation to always use the trampoline function.
Apple's pthread_setname_np() has a different calling convention and
a thread can only set its own name whereas on linux, a thread can
set any name and has a thread parameter to identify which thread.
We can adapt the linux code to work in much the same way as the
OSX code and reduce the amount of code under #ifdef to just the
pthread_setname_np() call.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>