Poll events and call advance_game() at 60Hz, and redraw the screen at 30Hz
(or redraw screen at 60Hz too if use_60_fps == 1) This fixes the "performance"
problem.  The problem was advance_game() has this hack to enable 60 FPS where
it expects to always be called at 60Hz, but every other time, it doesn't do
anything. If you call it at 30Hz, it only runs at 15Hz. That's where the
"performance" problem was coming from.  So now we poll events and call
advance_game() at 60Hz, like we're supposed to, and we redraw the screen
at either 30 Hz or 60 Hz depending on the value of use_60_fps.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>