Use -fno-common gcc flag
This means that globals declared in headers must be declared
in exactly one compilation unit without an "extern" storage class
and all declarations in other compilation units mush have the
"extern" storage class.  Gcc 10 makes -fno-compilation the default,
and we want to catch this problem on older compilers too, so we will
explicitly enable -fno-common.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>