Remove most remaining vestiges of gtk
There are still a few harmless vestiges around setting up colors,
which needed a gtk_widget which was passed as a void *, the opengl
code for this is a noop. So it could be deleted. But I think there's
nothing left that will cause the build process to emit warnings or
cause problems.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>