Use same near zero threshold for setting commodity thresholds as before
Recent change to allow lua scripts to empty player and npc cargo
bays by setting the index to -1 also changed the near zero threshold
on quantities from < 0.01 to <= 0.01.  This change wasn't intentional,
so I'm changing it back.  That being said, it's a pretty arbitrary
number, so I do not expect that this really affects anything.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>