snis_server: Set multiverse server port to -1 on failure.
Otherwise, it is possible that when snis_multiverse is restarted,
it might get the same IP and port as before, and then snis_server
won't recognize that it has come back from the dead, and won't
attempt to reconnect to it.  However, with this change, snis_server
might receive stale data from ssgl_server, with the old port number
of the old, dead snis_multiverse, and it will attempt to connect to
it and fail, because it's not there.  Eventually (within 10-20 secs),
ssgl_server will get the new port number and the old multiverse will
expire out of its list, so eventually, it should work, and snis_server
will connect to the new snis_multiverse, even if it has the same
port number as it did previously.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>