snis_client: Make pull down menus work on demon screen even when console is active
The button press and release code for the ui element list needs to find the *last*
candidate in the list to trigger, not the first, to match how things are drawn
(e.g. the last thing in the list is drawn last, or "on top of" everything before it.
And the thing "on top" should be the one to get the mouse events.)

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>