Blind attempt to re-debianize SDL changes
This is a blind (i.e. **untested**) attempt to fix #305.

Copy-pasting the package names from
https://github.com/smcameron/space-nerds-in-space#step-1-install-dependencies
and putting those in the right part of the `debian/control` file *should* be
enough.

Alas, I'm not in the best position right not to test if the SNIS `.deb`
builds/runs as before the GTK→SDL changes

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>