Fix bug that object selection on demon screen didn't always work correctly
Symptoms of the bug were that sometimes object selection appeared not to work
(would silently select the wrong object) and navigation on the demon screen
would sometimes seem to travel to the wrong location (again, wrong object
silently selected).

The cause of the bug was 1) that add_entity() did not initialize the sx, sy
or onscreen fields of struct entity, and 2) the code that checked the mouse
position vs. the entity's onscreen location was relying on the sx, sy coord
of the entity from the *previous frame*, which only happened to work most of
the time because the entity allocation would work out identically or
identically enough from one frame to the next.  On those occasions the entity
allocation worked out differently as may obviously happen, the wrong object
would usually be selected.

The fix was to initialize sx, sy and onscreen fields in add_entity(), and in
demon_show_3d(), move the call to render_entities() *before* the code that
checks the entity's onscreen position vs. the mouse coordinates rather than
after (because render_entities is what fills in sx, sy, and onscreen with
correct values for the current frame).

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>