Avoid crash due to race between SDL_DestroyWindow and SDL_PollEvent
The splash screen at start up of snis_client is created and destroyed
in its own thread.  SDL_PollEvent() is apparently not threadsafe inasmuch
as it doesn't seem to like SDL_DestroyWindow() getting called concurrently
from another thread (by which I mean it will occasionally core dump or
if it doesn't crash, X windows will complain about a "BadWindow" and then
it hangs.)

The workaround is just to call SDL_HideWindow() instead of SDL_DestroyWindow()
to avoid ripping the rug out from under SDL_PollEvent.

At this point, the above is just my theory about what's going wrong and I'm
not 100% sure the workaround will fix it, though I'm confident enough to
commit it. With this patch I wasn't so far able to reproduce the problem
anymore (although it's somewhat difficult to reproduce in any case.)

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>