Slightly improve obj file parsing
Example, if parsing a line that can be either like:

v 1.0 2.0 3.0 4.0
v 1.0 2.0 3.0

with an optional implicit 4th value of 1.0, we first tried to parse it via
sscanf() with "%f %f %f %f", and if that did not convert 4 items, we again
parse it with sscanf() with "%f %f %f".  We don't need to sscanf it again
if it converted 3 items the first time around.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>