snis_client: prevent infinite window resize event loop
If you are running wayland or a window manager other than X11
then the aspect ratio constraining code would resize the window
to maintain the aspect ratio, which caused a window resize event,
which caused the aspect ratio to be constrained, which resized
the window, which caused a window resize event, which... and so on.

We should check that the requested window size is different from
the current window size before resizing the window.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>