Fix TESTDOCK.LUA script
It did a clear_all() and creates a starbase at 1000,1000,1000, then
insta-docks the player to this starbase.  However, since the Sun is now
at 0,0,0, the location of the starbase is *way* too close to the sun,
and the player is killed immediately by the radiation/heat.

So, we need to move the starbase farther away from the sun.

Also, modify it to damage the player ship somewhat so we can
test the starbase repairing the player ship.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>