Fix bug in tsd.ship.damage_data_dirty flag usage
in player_move(), we had this:

        o->tsd.ship.damage_data_dirty = do_overheating_damage(o);
        o->tsd.ship.damage_data_dirty |= do_sunburn_damage(o);

damage_data_dirty gets cleared in exactly one place, when we add the data into the queues
to be pumped out to the clients.

However do_overheating_damage() can return zero, and thus clear damage_data_dirty
when it should not be cleared because the data hasn't yet been queued.  Should be:

        o->tsd.ship.damage_data_dirty |= do_overheating_damage(o);
        o->tsd.ship.damage_data_dirty |= do_sunburn_damage(o);

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>