Reduce the size of the large explosion spark
Explosions consist of many small sparks spreading out
from a central point, slowly shrinking.  But there is also
an initial *very* large spark that doesn't move and rapidly
shrinks to create a "flash" of light.  The size of this large
spark was absurdly large.  Reduced it by a factor of 5 and I
think it's about right now.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>