Force tonemapping limit constants to floats
Rather than allowing them to defaulti to doubles.  This caused a bug which made
the range checking done in tweak_variable() (which is all floats) subtly
different than the range checking in graph_dev_set_tonemapping_gain() which was
using the constants from snis_graph.h, which were doubles.  This meant you
could tweak tonemapping_gain to 2.2 via tweak_variable(), but when
graph_dev_set_tonemapping_gain() was called, it rejected 2.2 as being outside
the range, because apparently (2.2f <= 2.2) evaluates to false.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>