Allow user to set VSYNC mode
Default is now vsync on (previously it was vsync off). I think
helps with screen tearing at the cost of a dropped frame.  There
is a user tweakable variable, "VSYNC" which can be set to -1, 0, or 1
to use "adaptive vsync", "vsync off", or "vsync on".  I found adaptive
vsync aggravates tearing, vsync off usually doesn't tear much but does
sometimes a little bit, and vsync on doesn't tear, but may drop a
frame instead of tearing.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>