Add new "projector" role
The "onscreen" opcode toggled the display mode of stations
which supported MAINSCREEN between a specified mode and the
MAINSCREEN mode. This unnecessarily connects the notion of
*supporting* the MAINSCREEN mode with *being* the main screen
(i.e. being the giant screen in front of everyone.)

This meant if people select mainscreen role on their individual
stations, they are subjected to it bizarrely switching to
another screen whenever someone else presses ctrl-O on another
station to project their screen on *the* main screen, and
then subjected to their station switching to MAINSCREEN when
that other person "unprojects" their screen by pressing ctrl-O
a second time.  This provides a way for the game to distinguish
between which station(s) is/are *actually* the main screen(s),
and those which merely support the main screen mode.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>