Use EINTR, not -EINTR
This fixes *very* old and *very* uncommon bugs (not sure I have
ever seen these bugs manifest).  The source of the bugs is that I used
to do linux drivers for a living, and inside the kernel, you check for
-EINTR, not EINTR as you do in userland, so my habit from work when I
wrote the code checking for EINTR was to check for -EINTR.  D'oh.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>