snis_client: clear depth buffer between nav and main rendering
If you "set main_nav_hybrid=1" on the console, the NAV screen first
draws the main screen, then draws the nav screen "on top of it."  Except,
some of the 3D elements of the nav screen sometimes can go *behind* some of the
main screen elements, which should not be the case.  We need to clear
the depth buffer after drawing the main screen.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>