snis_multiverse: unignore SIGCHLD
When you start snis_multiverse via snis_launcher, SIGCHLD is not ignored.
When you start snis_multiverse vis snis_client, SIGCHILD is ignored, because
it inherits all the sigactions from snis_client.  We want to unignore this
signal in snis_multiverse, so it can get the status of the process when
starting up snis_servers.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>