ssgl_server: hand out two IP addresses for each server
First the one the server tells us, and second the remote_addr we
got back from accept.  These can differ.  If ssgl_server and snis_server
are on the same, host on a lan behind NAT, one of the addresses will be
127.0.0.1, and the other will be like 192.168.x.x.  If ssgl is on the
internet and snis_server is on a LAN behind NAT, one address will be
192.168.x.x or similar, the other will be the IP address of the
external router.

With port forwarding, this should make things work better over the
internet.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>