Implement NAT ghetto mode in snis_server and snis_client
If the first two connection attempts from snis_server to multiverse or from
snis_client to snis_server fail with the IP addresses obtained from ssgl_server
try using the lobby host IP address with the port number from ssgl_server.

The assumption is that most likely people are running ssgl_server, snis_server,
and multiverse all on the same host.  NAT has probably prevented ssgl_server from
obtaining the correct IP addresses.  But we already connected to ssgl_server, so
if ssgl_server is in the same host as the other two, maybe try that IP address
for them.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>